part way through refactor
